+++ aiidalab_qe/app/common/widgets.py
