+++ aiidalab_qe/app/configure/advanced.py
@@ -3,9 +3,9 @@ from aiida.plugins import DataFactory
 from aiida_quantumespresso.workflows.pw.base import PwBaseWorkChain
 from IPython.display import clear_output, display
 
-from aiidalab_qe.app.panel import Panel
+from aiidalab_qe.app.common.panel import Panel
 from aiidalab_qe.app.parameters import DEFAULT_PARAMETERS as QEAPP_DEFAULT_PARAMETERS
-from aiidalab_qe.app.pseudos import PseudoFamilySelector
+from aiidalab_qe.app.configure.pseudos import PseudoFamilySelector
 
 StructureData = DataFactory("core.structure")
