+++ aiidalab_qe/app/configure/workflow.py
@@ -1,7 +1,7 @@
 import ipywidgets as ipw
 
-from aiidalab_qe.app.panel import Panel
-from aiidalab_qe.app.utils import get_entry_items
+from aiidalab_qe.app.common.panel import Panel
+from aiidalab_qe.app.common.utils import get_entry_items
 
 
 class WorkChainSettings(Panel):
