+++ aiidalab_qe/app/result/node_view.py
@@ -21,7 +21,7 @@ from jinja2 import Environment
 
 from aiidalab_qe.app import static
 from aiidalab_qe.app.plugins.relax.result import Result as RelaxResult
-from aiidalab_qe.app.utils import get_entry_items
+from aiidalab_qe.app.common.utils import get_entry_items
 
 # I removed this `MinimalStructureViewer` class, because it is not used anywhere
