+++ tests/test_sssp.py
@@ -1,6 +1,6 @@
 def test_sssp(sssp):
     """Test the setup of SSSP pseudopotentials."""
-    from aiidalab_qe.app.sssp import install, install_pseudos, pseudos_to_install
+    from aiidalab_qe.app.submit.sssp import install, install_pseudos, pseudos_to_install
 
     assert len(pseudos_to_install()) == 0
     #
@@ -11,7 +11,7 @@ def test_sssp(sssp):
 
 def test_sssp_widget(sssp):
     """Test the widget for installing SSSP pseudopotentials.""" ""
-    from aiidalab_qe.app.sssp import SSSPInstallWidget
+    from aiidalab_qe.app.submit.sssp import SSSPInstallWidget
 
     widget = SSSPInstallWidget(auto_start=False)
     widget.set_message("abc")
